qd2/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2

# Vector database
qdrant-client==1.11.3

# Embeddings and LLM
openai==1.47.0

# Database connectors
#psycopg2-binary==2.9.9
pymongo==4.8.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
httpx==0.27.2

# Logging and monitoring
structlog==24.4.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0

